script/pyproject.toml
[tool.poetry]
name = "genai4fuzz"
version = "0.1.0"
description = ""
authors = ["Fausto Carvalho <faustocarva@gmail.com>"]
readme = "README.md"
packages = [{include = "genai4fuzz"}]

[tool.poetry.dependencies]
python = "^3.10"
fire = "^0.5.0"
openai = "^1.26.0"
slither-analyzer = "0.9.6"
tiktoken = "^0.6.0"
langchain-openai = "^0.1.6"
loguru = "^0.7.2"
web3 = "^6.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
genai4fuzz = "genai4fuzz:main"
